# internal/app/openapi.yaml
openapi: 3.1.0
info:
  title: WaCalls API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: >-
    HTTP API for WaCalls, a self-hosted WhatsApp calling bridge. All /api/*
    endpoints require the bearer token configured via WACALLS_API_TOKEN
    (when unset, authentication is disabled). The SSE stream also accepts the
    token as the access_token query parameter. This document is the canonical
    contract: a test gate fails the build when routes and spec drift apart.
servers:
  - url: /
security:
  - bearerAuth: []
  - queryToken: []
paths:
...
  /api/sessions/{sid}/calls:
    post:
      operationId: startCall
      summary: Start an outgoing call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/ClientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  description: E.164-ish phone number; non-digits are stripped
                video:
                  type: boolean
                  description: >-
                    Place the call as a video call (advertises video in the offer).
                    Video media is not yet carried, so the peer sees no picture until
                    the media legs land; signaling only.
      responses:
        "200": { $ref: "#/components/responses/CallRef" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Operator identified by X-Client-Id is already on a call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Per-session concurrent call cap reached, or rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500": { $ref: "#/components/responses/InternalError" }
        "503":
          description: Session is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listCalls
      summary: Live calls of the session, same shape as the SSE call-list
      parameters:
        - $ref: "#/components/parameters/Sid"
      responses:
        "200":
          description: Live calls ordered by startedAt then callId
          content:
            application/json:
              schema:
                type: object
                required: [calls]
                properties:
                  calls:
                    type: array
                    items:
                      $ref: "#/components/schemas/CallRecord"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/calls/{id}:
    get:
      operationId: getCall
      summary: One live call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          description: The live call; 404 once it ends or if it belongs to another session
          content:
            application/json:
              schema:
                type: object
                required: [call]
                properties:
                  call:
                    $ref: "#/components/schemas/CallRecord"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
    delete:
      operationId: endCall
      summary: End an active call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "204":
          description: Call ended (idempotent)
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /api/sessions/{sid}/calls/{id}/accept:
    post:
      operationId: acceptCall
      summary: Accept an incoming call, claiming it for the X-Client-Id operator
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200": { $ref: "#/components/responses/CallRef" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Claimed by another client, or operator already on a call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/calls/{id}/reject:
    post:
      operationId: rejectCall
      summary: Reject an incoming call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          description: Rejected (idempotent when the call is already gone)
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Call is in a state that cannot be rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /api/sessions/{sid}/calls/{id}/video:
    post:
      operationId: videoAction
      summary: Drive the call's video flow (upgrade, accept, reject, stop, orientation)
      description: >-
        Signals a video-flow transition to the peer. "request" asks the peer to add
        video, "accept"/"reject" answer the peer's request, "stop" turns local video
        off, and "orientation" reports the local device orientation (0..3). Video
        media is not yet carried; this endpoint only drives the signaling state machine.
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [request, accept, reject, stop, orientation]
                orientation:
                  type: integer
                  minimum: 0
                  maximum: 3
                  description: Required when action is "orientation".
      responses:
        "200":
          description: Transition signaled
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "400":
          description: Missing or invalid action/orientation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Call is not in a state that accepts the transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    queryToken:
      type: apiKey
      in: query
      name: access_token
  parameters:
    Sid:
      name: sid
      in: path
      required: true
      schema:
        type: string
      description: Session id
    CallId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Call id
    ClientId:
      name: X-Client-Id
      in: header
      required: false
      schema:
        type: string
      description: Operator identity; omitted means no ownership is recorded
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Per-IP rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such session or call
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid request parameter or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    CallRef:
      description: Reference to the affected call
      content:
        application/json:
          schema:
            type: object
            required: [call]
            properties:
              call:
                type: object
                required: [callId]
                properties:
                  callId:
                    type: string
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CallStatus:
      type: string
      enum: [starting, ringing, connected, reconnecting, ended]
...
    CallRecord:
      type: object
      required: [sessionId, callId, owner, direction, peer, startedAt, status]
      properties:
        sessionId:
          type: string
        callId:
          type: string
        owner:
          type: [string, "null"]
          description: Claiming operator's client id; null when unclaimed, never ""
        direction:
          type: string
          enum: [outbound, inbound]
        peer:
          type: string
        peerName:
          type: string
          description: Resolved contact name for the peer, when known
        peerPhotoUrl:
          type: string
          description: Cached WhatsApp profile-picture URL for the peer, when known
        startedAt:
          type: integer
          format: int64
          description: Unix milliseconds
        status:
          $ref: "#/components/schemas/CallStatus"
        video:
          type: boolean
          description: Whether the call was offered as a video call; absent for audio
        endedAt:
          type: integer
          format: int64
          description: Unix milliseconds; absent while the call is live
        endReason:
          type: string
          description: Absent while the call is live
...